criação dos testes respectivos a classe Atividade Alterações nas classes dependentes para funcionamento do programa

+++ Programacao3/src/trabalho/view/AtividadeView.java
@@ -25,7 +25,7 @@ public class AtividadeView {
 			System.out.println("\n data: ");
 			String data = ler.nextLine();
 			System.out.println("\n Valor: ");
-			float valor = sc.nextFloat();
+			Double valor = sc.nextDouble();
 			newAtividade = new Atividade(nome,tipo,data,valor);
 			this.atividades.salvar(newAtividade);
